requirements.txt
pynput
pillow
psutil
sounddevice
scipy
requests